# gravitee-apim-rest-api/gravitee-apim-rest-api-portal/gravitee-apim-rest-api-portal-rest/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Gravitee.io Portal Rest API
  description: API dedicated to the devportal part of Gravitee
  termsOfService: terms
  contact:
    name: The Gravitee Team
    url: https://gravitee.io
    email: contact@graviteesource.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: "3.10.9-SNAPSHOT"
servers:
  - url: http://localhost:8083/portal/environments/{envId}
    description: The portal API for a given environment
    variables:
      envId:
        default: DEFAULT
        description: ID of the environment
...
security:
  - BasicAuth: []
  - CookieAuth: []
paths:
...
  /applications/{applicationId}:
    parameters:
      - $ref: '#/components/parameters/applicationIdParam'
    get:
      tags:
        - Application
      summary: Get an application.
      description: |
        Get an application.

        User must have the APPLICATION_DEFINITION[READ] permission.
      operationId: getApplicationByApplicationId
      responses:
        200:
          description: One application.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          $ref: '#/components/responses/ApplicationNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Application
      requestBody:
        description: Use to update an application.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Application'
      summary: Update an application.
      description: |
        Update an application.

        User must have APPLICATION_DEFINITION[UPDATE] permission.
      operationId: updateApplicationByApplicationId
      responses:
        200:
          description: Updated application.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        400:
          description: Bad Request if body is empty.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          $ref: '#/components/responses/ApplicationNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Application
      summary: Delete an application
      description: |
        Delete an application.

        User must have the APPLICATION_DEFINITION[DELETE] permission.
      operationId: deleteApplicationByApplicationId
      responses:
        204:
          description: No-Content
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          $ref: '#/components/responses/ApplicationNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
  /applications/{applicationId}/picture:
    parameters:
      - $ref: '#/components/parameters/applicationIdParam'
    get:
      tags:
        - Application
      summary: Get the application's picture
      description: |
        Get the application's picture.

        User must have APPLICATION_DEFINITION[READ] permission.
      operationId: getApplicationPictureByApplicationId
      responses:
        200:
          $ref: '#/components/responses/PictureSuccess'
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          $ref: '#/components/responses/ApplicationNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
  /applications/{applicationId}/background:
    parameters:
      - $ref: '#/components/parameters/applicationIdParam'
    get:
      tags:
        - Application
      summary: Get the application's background
      description: |
        Get the application's background.

        User must have APPLICATION_DEFINITION[READ] permission.
      operationId: getApplicationBackgroundByApplicationId
      responses:
        200:
          $ref: '#/components/responses/PictureSuccess'
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          $ref: '#/components/responses/ApplicationNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
  /applications/{applicationId}/alerts:
    parameters:
      - $ref: '#/components/parameters/applicationIdParam'
    get:
      tags:
        - Application
      summary: Get application alerts
      description: |
        Get application alerts.

        User must have APPLICATION_ALERT[READ] permission to get alerts.
      operationId: getAlertsByApplicationId
      responses:
        200:
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alert'
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          $ref: '#/components/responses/ApplicationNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Application
      requestBody:
        description: Use to create a post.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertInput'
      summary: Create an application alert
      description: |
        Create an application alert.

        User must have the APPLICATION_ALERT[CREATE] permission.
      operationId: createApplicationAlert
      responses:
        201:
          description: Created alert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          $ref: '#/components/responses/ApplicationNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /applications/{applicationId}/notifications:
    parameters:
      - $ref: '#/components/parameters/applicationIdParam'
    get:
      tags:
        - Application
      summary: Get application notifications
      description: |
        Get application notifications.

        User must have APPLICATION_NOTIFICATION[READ] permission to get notifications.
      operationId: getNotificationsByApplicationId
      responses:
        200:
          description: List of notifications configured
          content:
            application/json:
              schema:
                description: List of notifications configured on the application.
                type: array
                items:
                  type: string
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          $ref: '#/components/responses/ApplicationNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Application
      requestBody:
        description: Use to update application notifications.
        content:
          application/json:
            schema:
              description: List of notifications configured on the application.
              $ref: '#/components/schemas/NotificationInput'
      summary: Update notifications for an application.
      description: |
        Update notifications for an application.

        User must have APPLICATION_NOTIFICATION[UPDATE] permission to update notifications.
      operationId: updateApplicationNotifications
      responses:
        200:
          description: Updated application notifications.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          $ref: '#/components/responses/ApplicationNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
  /applications/{applicationId}/members:
    parameters:
      - $ref: '#/components/parameters/applicationIdParam'
    get:
      tags:
        - Application
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
      summary: List application members
      description: |
        List application members.

        User must have the APPLICATION_MEMBER[READ] permission.
      operationId: getMembersByApplicationId
      responses:
        200:
          description: List of members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembersResponse'
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          $ref: '#/components/responses/ApplicationNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Application
      requestBody:
        description: Use to create a member.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberInput'
      summary: Create an application member
      description: |
        Create an application member.

        User must have the APPLICATION_MEMBER[CREATE] permission.
      operationId: createApplicationMember
      responses:
        201:
          description: Created member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          $ref: '#/components/responses/ApplicationNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /applications/{applicationId}/metadata:
    parameters:
      - $ref: '#/components/parameters/applicationIdParam'
    get:
      tags:
        - Application
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
      summary: List application metadata
      description: |
        List application metadata.

        User must have the APPLICATION_METADATA[READ] permission.
      operationId: getMetadataByApplicationId
      responses:
        200:
          description: List of metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceMetadataResponse'
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          $ref: '#/components/responses/ApplicationNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Application
      requestBody:
        description: Use to create a metadata.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReferenceMetadataInput'
      summary: Create an application metadata
      description: |
        Create an application metadata.

        User must have the APPLICATION_METADATA[CREATE] permission.
      operationId: createApplicationMetadata
      responses:
        201:
          description: Created metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceMetadata'
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          $ref: '#/components/responses/ApplicationNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /applications/{applicationId}/analytics:
    parameters:
      - $ref: '#/components/parameters/applicationIdParam'
    get:
      tags:
        - Application
        - Analytics
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
        - $ref: '#/components/parameters/fromQueryParam'
        - $ref: '#/components/parameters/toQueryParam'
        - $ref: '#/components/parameters/intervalQueryParam'
        - $ref: '#/components/parameters/queryQueryParam'
        - $ref: '#/components/parameters/fieldQueryParam'
        - $ref: '#/components/parameters/analyticTypeQueryParam'
        - $ref: '#/components/parameters/analyticRangeQueryParam'
        - $ref: '#/components/parameters/analyticAggregationQueryParam'
        - $ref: '#/components/parameters/analyticOrderQueryParam'
      summary: Get Application analytics
      description: |
        Get the application analytics.

        User must have the APPLICATION_ANALYTICS[READ] permission.
      operationId: getApplicationAnalytics
      responses:
        200:
          description: Analytics
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/DateHistoAnalytics'
                  - $ref: '#/components/schemas/GroupByAnalytics'
                  - $ref: '#/components/schemas/CountAnalytics'
        400:
          description: Invalid query params.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          $ref: '#/components/responses/ApplicationNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
  /applications/{applicationId}/configuration:
    parameters:
      - $ref: '#/components/parameters/applicationIdParam'
    get:
      tags:
        - Application
      summary: Get the application type configuration.
      operationId: getApplicationType
      description: |
        Get application type.
      security: []
      responses:
        200:
          description: Application type configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationType'
        500:
          $ref: '#/components/responses/InternalServerError'
  /applications/{applicationId}/logs:
    parameters:
      - $ref: '#/components/parameters/applicationIdParam'
    get:
      tags:
        - Application
        - Analytics
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
        - $ref: '#/components/parameters/fromQueryParam'
        - $ref: '#/components/parameters/toQueryParam'
        - $ref: '#/components/parameters/queryQueryParam'
        - $ref: '#/components/parameters/fieldQueryParam'
        - $ref: '#/components/parameters/logOrderQueryParam'
      summary: Get Application logs
      description: |
        Get the application logs.

        User must have the APPLICATION_LOG[READ] permission.
      operationId: getApplicationLogs
      responses:
        200:
          description: List of logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogsResponse'
        400:
          description: Invalid query params.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          $ref: '#/components/responses/ApplicationNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /applications/{applicationId}/_renew_secret:
    parameters:
      - $ref: '#/components/parameters/applicationIdParam'
    post:
      tags:
        - Application
      summary: Renew the client secret for an OAuth2 application
      description: |
        Renew the client secret for an OAuth2 application.

        User must have the APPLICATION_DEFINITION[UPDATE] permission.
      operationId: renewApplicationSecret
      responses:
        200:
          description: Updated application with a new secret.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        403:
          $ref: '#/components/responses/PermissionError'
        404:
          $ref: '#/components/responses/ApplicationNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
  /applications/{applicationId}/subscribers:
    parameters:
      - $ref: '#/components/parameters/applicationIdParam'
    get:
      tags:
        - Application
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
        - $ref: '#/components/parameters/subscriptionStatusesQueryParam'

      summary: List APIs that subscribed with an application
      description: |
        Lists API that current user is allowed to access. May be filtered by status. Ordered by nimber of hits.

        This application has to be accessible by the current user, otherwise a 404 will be returned.
      operationId: getSubscriberApisByApplicationId
      security: []
      responses:
        200:
          description: List accessible APIs for current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApisResponse'
        404:
          $ref: '#/components/responses/ApplicationNotFoundError'
        500:
          $ref: '#/components/responses/InternalServerError'
  /groups:
    get:
      tags:
        - Group
      parameters:
        - $ref: '#/components/parameters/pageNumberParam'
        - $ref: '#/components/parameters/pageSizeParam'
      summary: List groups
      description: |
        List all groups

        User must have the MANAGEMENT_GROUP[READ] permission.
      operationId: getGroups
      responses:
        200:
          description: List of groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupsResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /configuration:
    get:
      tags:
        - Portal
      summary: Get portal configuration.
      operationId: getPortalConfiguration
      description: |
        Get all the portal configuration from the platform settings.
      security: []
      responses:
        200:
          description: Portal configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationResponse'
        500:
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  parameters:
...
    applicationIdParam:
      name: applicationId
      in: path
      required: true
      description: Id of an application.
      schema:
        type: string
...
    pageNumberParam:
      name: page
      in: query
      required: false
      description: The page number for pagination.
      schema:
        type: integer
        default: 1
    pageSizeParam:
      name: size
      in: query
      required: false
      description: |
        The number of items per page for pagination.
        If the size is 0, the response contains only metadata and returns the values as for a non-paged resource.
        If the size is -1, the response contains all datas.
      schema:
        type: integer
        default: 10

    #Apis filtering
...
    fromQueryParam:
      name: from
      in: query
      required: false
      description: Lower bound of timestamp for filtering.
      schema:
        type: integer
        format: int64
...
  schemas:
...
    ErrorResponse:
      properties:
        errors:
          description: List of errors
          type: array
          items:
            $ref: '#/components/schemas/Error'
...
    MembersResponse:
      properties:
        data:
          description: List of members.
          type: array
          items:
            $ref: '#/components/schemas/Member'
        metadata:
          $ref: '#/components/schemas/MetadataMap'
        links:
          $ref: '#/components/schemas/Links'
...
    Application:
      properties:
        id:
          description: Unique identifier of an application.
          type: string
        name:
          description: Name of the application.
          type: string
        description:
          description: Description of the application.
          type: string
        applicationType:
          description: Type of the application (Web, Mobile, ...).
          type: string
        hasClientId:
          description:  True if the application has client id.
          type: boolean
        owner:
          $ref: '#/components/schemas/User'
        created_at:
          description: Creation date and time of the application.
          type: string
          format: date-time
        updated_at:
          description: Last update date and time of the application.
          type: string
          format: date-time
        groups:
          description: Array of groups associated to the application.
          type: array
          items:
            $ref: '#/components/schemas/Group'
        picture:
          description: |
            Picture of the application. This attribute is only used to update a picture.\
            To get the application picture, use /application/{applicationId}/picture.
          type: string
          writeOnly: true
        background:
          description: |
            Background of the application. This attribute is only used to update a picture.\
            To get the application picture, use /application/{applicationId}/background.
          type: string
          writeOnly: true
        settings:
          $ref: '#/components/schemas/ApplicationSettings'
        _links:
          $ref: '#/components/schemas/ApplicationLinks'
...
    Alert:
      properties:
        id:
          description: Unique identifier of an alert.
          type: string
        enabled:
          description: true, if alert is enabled
          type: boolean
        type:
          $ref: '#/components/schemas/AlertType'
        status_code:
          description: Http status code to trigger the alert
          type: string
        status_percent:
          description: Percent to trigger the alert on status code
          type: integer
          minimum: 1
          maximum: 100
        response_time:
          description: Response time to trigger the alert
          type: integer
          minimum: 1
          maximum: 100000
        duration:
          description: Compute alert on selected duration
          type: integer
        time_unit:
          $ref: '#/components/schemas/AlertTimeUnit'
    Member:
      properties:
        id:
          description: Unique identifier of a member (user).
          type: string
        user:
          $ref: '#/components/schemas/User'
        created_at:
          description: Creation date and time of the member.
          type: string
          format: date-time
        updated_at:
          description: Last update date and time of the member.
          type: string
          format: date-time
        role:
          description: Role of the member. (OWNER, USER, ...).
          type: string
...
    MemberInput:
      properties:
        user:
          type: string
          description: User's technical identifier.
        reference:
          type: string
          description: User's reference for user providing from an identity provider.
        role:
          type: string
          description: Role's name
    AlertInput:
      properties:
        type:
          $ref: '#/components/schemas/AlertType'
        status_code:
          description: Http status code to trigger the alert
          type: string
        status_percent:
          description: Percent to trigger the alert on status code
          type: integer
          minimum: 1
          maximum: 100
        enabled:
          description: true, if alert is enabled
          type: boolean
          default: true
        response_time:
          description: Response time to trigger the alert
          type: integer
          minimum: 1
          maximum: 100000
        duration:
          description: Compute alert on selected duration
          type: integer
        time_unit:
          $ref: '#/components/schemas/AlertTimeUnit'
...
    NotificationInput:
      properties:
        hooks:
          type: array
          items:
            type: string

    #######################
    # Composition Objects #
    #######################
...
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PermissionError:
      description: Insufficient rights to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ApplicationNotFoundError:
      description: Application not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    PictureSuccess:
      description: A picture
      content:
        'image/*':
          schema:
            type: string
            format: binary
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    CookieAuth:
      type: apiKey
      in: cookie
      name: Auth-Graviteeio-APIM